Fix building from a clean git directory.
The file 'wfdblib.h' (along with other files that are created or
modified while building the library) has been removed from the git
repository to avoid polluting diffs.  However, this caused 'make' or
'make install' from the toplevel directory to fail unless you first
ran 'make' or 'make setup' in the lib subdirectory.

As a side effect, this means that the version of wfdblib.h installed
in $(INCDIR)/wfdb will match the version used to build the library.
(I'm not convinced that wfdblib.h *should* be installed at all, but if
it is installed, it should at least be the correct version.)