Fix password-leaking issues in pnwlogin.
Previously, 'pnwlogin' was implemented by setting the PNWUSER and
PNWPASS environment variables; those variables were passed to libcurl
unconditionally, and it would use that username and password for all
network transfers (provided that the server requested authentication.)
This was problematic due to the possibility of a malicious or
misconfigured web server or proxy somewhere; but more to the point, it
was a serious problem because libcurl would remember that
physionet.org requested Basic authentication over HTTPS - at which
point, it would also include a Basic authentication header with every
HTTP request to physionet.org.  (This is arguably a bug in libcurl.)

As a result, anyone who used 'pnwlogin' to access protected databases
(unless they manually removed http://physionet.org/physiobank/database
from their WFDB path) would have exposed their username and password
to any passive attacker snooping on the network.

The new implementation uses a completely different environment
variable (WFDBPASSWORD) which limits password authentication to a
particular protocol, server, and optionally a particular subdirectory
on that server.  WFDBPASSWORD may also contain multiple entries, in
order to make it easier to switch between multiple protected data
servers.  (Using multiple protected data servers simultaneously may
have unknown security implications and is not recommended.)

The new version of pnwlogin deliberately does not save the PNWUSER and
PNWPASS variables in the environment, in case a user tries to run a
program linked with an old, broken version of libwfdb.

Note also that the old version of pnwlogin does not work anymore; the
PhysioNet server will reject login attempts that do not include the
query string 'pnwlogin=10.5.24'.