Use different LDFLAGS for static versus dynamic libraries.
If LDFLAGS is specified in site-slib.def, and we are building a
dynamic library, copy those flags over to site.def.  (Note that,
confusingly, site-slib.def is used only for building the dynamic
libwfdb, whereas site.def is used for building both the static libwfdb
and the applications.)

This is a little kludgy, but will allow us to use the correct LDFLAGS
in all cases, since often the LDFLAGS needed for a static library are
not the same as those needed for a dynamic library.