Various fixes to LDFLAGS.
On Cygwin, Darwin, FreeBSD, Linux, MinGW, do not include $(LL) in
LDFLAGS when linking to a shared libwfdb.  On those platforms, libwfdb
is linked with libcurl if necessary; it is not necessary or
particularly helpful for applications to be linked with libcurl as
well.  (The same may also be true of Solaris and HP-UX, but I'm not
sure.)

On Darwin, FreeBSD, and generic, when building libwfdb itself, place
$(LL) in the correct position in the command line (after the list of
object files.)  I don't know if this is actually necessary or not, but
it won't hurt.  It's necessary if the linker works from left to right
and uses "--as-needed"-like behavior by default.

On FreeBSD (static libwfdb) and Solaris (shared libwfdb), use the
correct variable ($(LL) rather than $(LDLIBWWW) $(LDLIBCURL).)