Define RPATHFLAGS for various systems.
This variable needs to be defined in both foo.def and foo-slib.def,
but it'll only actually be used if we are building a shared library,
so it should only be mentioned in LDFLAGS in foo-slib.def.

These should work on Linux and FreeBSD at least.  Solaris is a shot in
the dark, as is HP-UX.  Darwin seems not to support RPATH but maybe
doesn't need it?  MinGW and Cygwin of course have nothing like RPATH.

Also, on Linux, do not use --no-as-needed.