Use /usr/local/lib as default LIBDIR.
In general, use $(WFDBROOT)/lib as LIBDIR unless the user manually
specifies --libdir to configure.  If the user specifies '-m32', use
$(WFDBROOT)/lib32 instead.

However, if '$(WFDBROOT)/lib64' is a standard library search directory
for the native architecture (as determined by the C compiler), then
use $(WFDBROOT)/lib64 for the native architecture, and $(WFDBROOT)/lib
for -m32.  (The effect of this is that, on Fedora, if you specify
--prefix=/usr, you will get the historical behavior.)

There are a few points to be made here:

 - /usr/local/lib is widely accepted, by basically everyone except Red
   Hat, as a standard location for installing libraries for the native
   architecture (regardless of what that architecture might be.)  It
   is also in the default ld.so search path on many systems, so on
   those systems 10.5.24 can be backwards compatible with 10.5.22, and
   programs can potentially avoid using RPATH.

 - /usr/local/lib64 is not a standard location according to anybody
   (not even Red Hat.)

 - Nowadays the only reason most people need a 32-bit libwfdb at all
   is for WAVE.  Since WAVE is built as part of the WFDB package we
   can ensure it gets built with the proper RPATH if necessary.

 - Backwards compatibility with 10.5.23 (in terms of RPATHs) is mostly
   a non-issue, since wfdb-config was unusable in 10.5.23.