Fix output of wfdb-config --libs.
The LDFLAGS variable is precisely what we want here, since the
Makefiles are designed to recompile the library and applications when
'make install' is run.

However, in some cases we need to include the curl library flags as
well as those for wfdb itself, so LDFLAGS may include backquoted
commands to be run.  Ensure that wfdb-config actually executes those
commands by passing wfdbldflags() to the shell.

Fixes #1 on github.