Fixes for cross-compiling.
Do not undefine _WIN32 as it breaks mingw system header files (and
probably other stuff too).  It's just a bad idea.

(In addition, note that wfdb.h itself was broken if _WIN32 was
defined; apparently nobody has tested that configuration in many
years.  For the time being I have left that code in place, guarded by
'#ifdef _WINDLL' instead of _WIN32.)

Ensure that we create BINDIR before copying the DLL there.  This
matters because now 'make all' sets LIBDIR to build/lib, whereas it
was previously set to build/bin (note that the nonstandard behavior of
installing .a files in 'bin' should probably be fixed as well.)

Please note that the package still does not officially support
cross-compiling.  If you are trying to cross-compile WFDB, you are on
your own.  Hopefully these changes are a step in the right direction,
however.