put_ann_table: fix output when both types of metadata are required.
If we are using both a custom time resolution (e.g. due to setafreq())
and custom annotation types (e.g. due to setannstr()), previous
versions of the library would add a NOTQRS following the 'time
resolution' comment and before the list of annotation types.  This
caused get_ann_table to stop parsing at that point, thus the
annotation types would not be loaded correctly and would appear to the
application as NOTE annotations.

The problem is due to the fact that putann() will write a 'time
resolution' comment itself - followed by a NOTQRS - if it sees that a
custom afreq has been defined.  This is a mess and I don't know if any
applications rely on this behavior (i.e. calling setafreq after
annopen rather than beforehand.)