Update Fortran wrappers for compatibility with modern systems.
As far as I can tell, all modern Fortran compilers will normally use
the C 'int' type as the default integer type.  If you are using a
Fortran compiler that uses 'long' as the default (for example,
gfortran on x86-64 with the -fdefault-integer-8 option), or some even
more exotic type, then you will need to specify that by defining the
INTEGER macro when compiling wfdbf.c.  Likewise you can define
DOUBLE_PRECISION to some type other than the standard 'double', if
necessary.

If the INTEGER type is not the same size as a WFDB_Sample, then arrays
passed to getvec, getframe, and putvec will be converted
appropriately.