configure: cross-compiling improvements.
Use --without-xview to avoid XView entirely, even if it appears to be
present.  Use --with-xview=DIR to explicitly specify the XView
installation prefix.

Furthermore, when cross-compiling, don't search for libraries
automatically (since what you will find are libraries for the build
system, which are unlikely to be useful.)  That is, if --host is
specified, it implies --without-netfiles and --without-xview, unless
--with-libcurl and/or --with-xview are also specified.