Support DESTDIR.
The DESTDIR variable is not normally defined.  If it is set on the
'make' command line, the package will be installed into subdirectories
of that directory (for example, if WFDBROOT=/usr/local and
DESTDIR=/tmp/wfdb, the package will be installed in
/tmp/wfdb/usr/local.)  However, unlike WFDBROOT, DESTDIR does not
affect built-in paths such as the default database path.

This allows building binary packages (such as deb and rpm packages)
without actually writing to the directories where the package will
eventually be installed.