wave: fix display when jumping back to start of record.
Previously, if we attempted to read a section of the signal that was
past the end of the file (isigsettime() returned an error), a display
list was allocated, and added to the cache, but never filled in with
valid data ('start' and 'ndpts' would be set to zero).

As a result, if we subsequently jumped back to the start of the
record, without first loading enough *other* display lists to flush
the invalid list out of the cache, nothing would be displayed.

The circumstances to trigger this bug were not obvious, but it was
most apparent with very short records.