wfdb_error, wfdb_open: fix a variety of buffer overflow issues.
WFDB applications could be made to crash or misbehave in a variety of
ways by supplying very long record or file names that would end up
being copied into 'wfdb_filename' or 'error_message'.

To fix these issues, define a pair of functions wfdb_vasprintf and
wfdb_asprintf (similar to GNU vasprintf and asprintf), and use these
functions throughout wfdb_error and wfdb_open.

In addition, in wfdb_open, correctly handle the corner case of a
record name that ends with a single-character directory name (for
example, the record "foo/a/", which is shorthand for "foo/a/a".)