lcheck: correctly check for library errors.
The previous code didn't do what was intended - although it copied the
result of wfdberror into a separate buffer, it would then call
wfdberror again and set libversion to point to the buffer that
wfdberror returned.  With the old version of wfdberror, this would
merely fail to detect errors (since it was comparing a string with
itself.)  With the new wfdberror, which dynamically allocates the
result, it would be accessing the buffer after it has been freed.