getann: handle huge time values, and round correctly.
The "raw" annotation time, previously stored internally as a
WFDB_Time, is now stored as a double.  This means that raw annotation
timestamps can now be larger than the maximum range of a WFDB_Time,
and the annotations will still be read correctly, so long as the
*scaled* timestamps fit within a WFDB_Time.  For example, if the
annotation time resolution is one millisecond, but the sampling
frequency is set to 250 Hz, 32-bit applications will be able to read
annotation files up to 8.5 million seconds (99 days) in length.

This does mean that if the raw annotation timestamps exceed the
precision of a double (typically 53 bits), there will be some
accumulated error over the course of a record.  (The scaled timestamps
returned by getann have always been limited to the precision of a
double, but on 64-bit systems, it was previously not possible for
rounding errors to accumulate.)

Additionally, the conversion from double to WFDB_Time now performs
correct rounding (for negative values), as well as bounds checking.
In particular, if the scaled annotation time is greater than the
maximum possible WFDB_Time, the maximum possible WFDB_Time will be
returned instead.  Previously, the result in that situation was
undefined behavior - often wrapping or returning a sentinel value such
as LONG_MIN, but unpredictable in general.