In high-resolution mode, set the sampling frequency in readheader.
This means that, following setgvmode(WFDB_HIGHRES), sampfreq("foo")
will return the highest sampling frequency of any signal in the record
"foo".  This is useful for reading and writing annotations and
performing time conversions for multi-frequency records, without
actually reading any signals.

On the other hand, when we call isigopen to open one or more signals,
the current sampling frequency will be set to match the getvec
frequency, which might be greater than the highest sampling frequency
in the record (if multiple records are opened at once) or might be
less (if not all signals are opened.)

This requires that we decouple the "current sampling frequency" and
"current frame frequency" from the "getvec conversion scale".

- sampfreq(), setsampfreq(), getspf(), [ms]timstr(), and strtim()
  (except strtim("i")) refer to the "current sampling frequency"
  (sfreq) and/or "current frame frequency" (ffreq).

- getvec(), isigsettime(), and strtim("i") refer to the "getvec
  conversion scale".

(It is still the case that calling sampfreq on an EDF record forces
the application into high-res mode if it wasn't already, so in that
regard sampfreq can affect the behavior of getvec.  This is arguably a
bug.)