When reading format 311, allow two samples packed into 3 bytes at EOF.
If the total number of samples in a signal file is 3*k, the signal
file is written as 4*k bytes.

If the total number of samples is 3*k+1, the signal file can be
written as 4*k+2 bytes.  Older versions of WFDB can read these files
correctly, but cannot write them correctly.

If the total number of samples is 3*k+2, then in principle, the file
only needs to be 4*k+3 bytes.  However, older versions of WFDB would
not be able to read such files.

(Another option is to pad the file to 4*k+4 bytes, but in that case,
if there is only one signal and one sample per frame, the record will
appear to be one sample longer, which may be undesirable.)

Note that reading format 212 has always worked correctly in this
regard.  Format 310 does not permit arbitrary numbers of samples.