Include leftover samples when flushing/closing a bit-packed output file.
When writing a signal file using format 212, 310, or 311, there may be
samples left over, if the total number of samples is not divisible by
2, 3, or 3, respectively.

When wfdbquit() is called, ensure that all of the required samples are
actually written to the output file before closing it.  (In some cases
this means padding the file with additional zero samples.)

When wfdbflush() is called, if the output file is seekable (i.e., it's
a regular file, rather than a pipe or something more exotic), and
block-sized output is not required, write out any leftover samples and
then seek back to the previous position.