setifreq: use nvsig, and correctly handle nvsig = 0.
Previously, setifreq would only allocate space for 'nisig' signals in
its internal buffer, not 'nvsig'; the two may not be equal if we are
reading a variable-layout record and setifreq is not called
immediately after isigopen.

Additionally, if nisig was zero, then the function would sometimes
work correctly and sometimes fail.  Specifically:

- On systems where realloc(0, 0) returns a null pointer, setifreq
  would always fail if no signals were open.

- On systems where realloc(0, 0) returns a non-null pointer (and
  realloc(realloc(0, 0), 0) returns null), setifreq would work the
  first time and fail the second time.