oannclose, sortann: When sorting a newly-created file, set the path to '.'.
Output annotation files are always (at least for now) created in the
current working directory.  However, that directory may or may not be
part of the WFDB path.  When oannclose invokes sortann, we always want
to sort the file that was just written, not some other file by the
same name located elsewhere.  We also want to avoid network access
that the application didn't ask for.

oannclose will now invoke sortann with '-r.' instead of '-r', and
sortann will recognize that as meaning the WFDB path should be set to
'.'.

(Other possible solutions: unconditionally setting the path would
break compatibility with old libraries/applications, and prevent
future versions from changing the output location; adding an entirely
new command-line option would break compatibility between the new
library and old binaries; calling putenv would muck up the calling
application's environment; setting the environment as part of the
'system' string would be dependent on the shell.)