annopen: defer writing annotation table until the first putann().
This means that an application may call setafreq(), setannstr(),
and/or setanndesc() either before or after annopen() - the value of
oafreq and the custom type strings take effect only when putann() is
called for the first time on a given output annotator.

Previously, the behavior was both awkward and inconsistent: setafreq()
could be called after annopen(), but only if there were no custom
annotation types defined before annopen().

This also means that if the application never calls putann(), the
resulting empty annotation file will not have any metadata either
(so an empty MIT-format annotation file will be exactly two bytes.)