New functions: setiafreq and getiafreq.
We now keep track of both the scaled and unscaled annotation times for
the currently buffered input annotation (ann), the pushed-back
annotation (pann), and the most recently returned annotation.

This permits the annotation time scale to be changed after an
annotation has already been read (or even after an annotation has been
returned by getann, then the same annotation pushed back by ungetann),
without loss of precision.

This also changes the behavior when calling annopen after setifreq.
In previous versions of the library, the frequency specified with
setifreq would only be respected if the annotation file included an
explicit time resolution.  That is to say, in previous versions, the
effective input time resolution was either sampfreq(NULL) or
getifreq(), and it was impossible for the application to know which.

Now, if setifreq is called before annopen, the specified resolution
will be used for both input and output annotations.

On the other hand, calling setifreq *after* annopen will resample
input signals without affecting input annotations; this behavior is
backwards-compatible.

Finally, we also avoid loss of precision in the case where a time
resolution comment is followed immediately by a normal annotation
(i.e., the expected NOTQRS is missing.)