mrgann: preserve original time resolution, and fix handling of -m.
If neither input file has an explicit resolution, the output file will
not have one either.

If both input files have explicit resolutions, the output file will
have the higher of the two resolutions.

If one input file has an explicit resolution and the other file does
not, then the latter is assumed to match the record frame frequency,
and the higher of those two resolutions is used.

In addition, fix a very old bug in the handling of -m options.  In
previous versions, the argument to the first -m option was parsed
*before* the sampling frequency was defined (meaning that argument was
usually interpreted as zero.)  This change brings the application's
behavior in line with its documentation.  For backwards compatibility,
when using -m, always begin with '-m0 0', '-m1 0', '-m2 0', or '-m3 0'.

The "expected" mrgann output has been updated to account for both of
these issues.