rdann: in -x mode, show higher than millisecond precision if possible.
If the resolution is less than 1000 ticks per second, use the
traditional format, with precision of .001 second in the "seconds"
column, .00001 minute (.6 ms) in the "minutes" column, and .0000001
hour (.3 ms) in the "hours" column.

If the resolution is greater than 1000 ticks per second, add
additional digits, as appropriate, to each column.

(This only applies to -x format because the other formats rely on
mstimstr, which is limited to millisecond precision.)