getann: allow ungetann to push back an annotation at time zero.
This was the behavior in older versions of WFDB, but was changed in
version 10.4.5 as part of the support for multi-frequency annotations.

In particular, this matters when reading an annotation file that does
not include a time resolution or annotation type definitions, because
get_ann_table will read ahead, and then push back the first annotation
it doesn't recognize.  In the case of mitdb record 220, for example,
the first reference annotation is a RHYTHM annotation at time zero.

Note that it is still not possible to push back a NOTQRS annotation
(and get_ann_table relies on this.)