wfdblib.h, wfdbio.c: provide stdio wrappers in non-NETFILES libraries.
When NETFILES is enabled, the library provides 11 functions that can
be used to read and write WFDB_FILEs:

  wfdb_clearerr (*)
  wfdb_feof
  wfdb_ferror
  wfdb_fflush
  wfdb_fgets
  wfdb_fread
  wfdb_fseek
  wfdb_ftell
  wfdb_fwrite
  wfdb_getc
  wfdb_putc

When using a non-NETFILES library, these functions are trivial
wrappers for the corresponding standard C functions.  Historically,
they have simply been defined as macros in wfdblib.h (except
wfdb_clearerr, which was mistakenly omitted.)  This was a problem
because:

- if an application (that uses these functions) was built using
  NETFILES-enabled headers and linked with a NETFILES-enabled shared
  library, it would then be unable to run with a non-NETFILES shared
  library;

- if an application was built using non-NETFILES headers and later run
  using a NETFILES shared library, it would crash if it tried to read
  a remote file.

Regardless of whether NETFILES is enabled, define these as external
functions, with prototypes in wfdblib.h, to provide ABI compatibility
both ways.