isigopen: use constant samples-per-frame in multi-segment records.
The layout header determines the ispfmax (number of samples per frame
in high-resolution mode), which is then kept constant for all other
segments (including gap segments as well as "real" segments).  This
means that ispfmax will not necessarily equal the maximum spf of any
open input signal.

This is necessary in order for getvec to behave sensibly when reading
multi-frequency, multi-segment records, particularly in
high-resolution mode.