isigopen: don't reallocate signal buffers needlessly.
This led to undefined behavior, for example, when reading a
multi-frequency variable-layout multi-segment record, because rgetvec
would call getframe(tvector), and getframe would call isigopen, and
isigopen would reallocate tvector.

Thus, although this change fixes the common case, there is still a lot
of serious breakage here - tvector, uvector, and vvector should only
be allocated by the functions that use them.  The larger problem is
the fact that buffers are reused both for *input* signals and
*virtual* signals; ideally, the library ought to support records with
framelen > vframelen and/or nisig > nvsig.