isgsetframe: disable checksum/EOF testing when seeking mid-byte.
When reading a bit-packed signal file and seeking to a position that
is not on a byte boundary, isgsetframe first calls itself recursively
(to seek to a byte boundary) and then calls r212/r310/r311 to skip
forward in the file.

This caused a problem in one particular corner case: when the
recursive call to isgsetframe short-circuited, it wouldn't disable
checksum/EOF testing.  (In fact, that behavior of isgsetframe is
intentional: getskewedframe calls isgsetframe at the beginning of the
record, which normally short-circuits, and in that instance we want
checksum testing to remain enabled.)

Consequently, for example, if a record has only one signal, with one
sample per frame, stored in format 212, 'rdsamp -f s1' would give an
error about an unexpected EOF and incorrect checksum.