isigsettime: correct behavior in high-resolution or ifreq mode.
istime is the next frame number to be returned by getframe.  t,
however, is a *sample* number, not a frame number.  In high-resolution
mode, the next sample number to be returned by rgetvec is (istime-1) *
ispfmax + gvc.  Thus, short-circuiting if t == istime is incorrect.

In *ifreq* mode, on the other hand, isigsettime(t) is expected to seek
to input sample number (WFDB_Time)(t * sfreq/ifreq), and restart the
resampling from that point.  For example, isigsettime(t-1) followed by
getvec is *not* equivalent to isigsettime(t) in most cases.  It's
complicated to figure out whether we can safely short-circuit in this
case... so just don't.