isgsetframe: fix handling of multi-group segments.
This has been badly broken in the past; for example, after calling
isigopen, isgsetframe would not even check whether g was still a valid
group number in the new segment.

It has never been possible to read separate groups from different
segments.  Moreover, in the variable-layout case, that wouldn't make
much sense, because as far as the application is concerned there is
only one signal group to begin with.

However, in the fixed-layout case, it may be possible to read separate
groups from different points in the same segment; leave that behavior
intact for now.