putann: time sanity checking.
It's (relatively) easy to make a mistake that results in calling
putann with time = WFDB_TIME_MIN or WFDB_TIME_MAX.  On a 64-bit
system, this can be a somewhat painful mistake.  So don't allow it in
general.

For consistency, forbid writing annotations at WFDB_TIME_MIN or
WFDB_TIME_MAX on all systems - it's "safe" enough on 32-bit systems,
but this has the benefit of ensuring that 32-bit mrgann, gqpost,
nguess, etc., will show an error if they are used on a record that is
too long for them to handle.