get_ann_table: slightly stricter parsing of annotation types.
An annotation type definition must actually begin with a decimal
number.  A comment such as "gqrs -r mitdb/100" should not be treated
as a definition for annotation type 0.

(I'm not requiring the number to constitute the entire first token, so
as to leave open the possibility of extending the format in the
future.)