wfdb2mat: version 5 output.
This is a very different, and generally inferior, format from the
"version 4" format used previously.  Its header is much more
complicated; it requires padding at the end of the file; and it is
limited to a maximum total size of 2^31 bytes.

However, Matlab arbitrarily refuses to read "version 4" files that
contain more than 100,000,000 elements.  So although it is possible
for a "version 4" file to contain as many as 2^31-1 rows and columns,
many applications are unable to read such files.

This patch was written by Chen Xie, with some minor edits from me.