xform: multifrequency fixes.
In -M mode, always set low-resolution mode initially, so that ifreq
and ofreq are set to the frame frequency, even for EDF; also so that
from, to, nsamp, and spm are calculated in frames.  But once that's
done, switch to high-resolution mode for annotations, and call
setafreq.

Note that the test suite involves extracting annotations from mimicdb,
and the previous version of the test suite "expected" the resulting
annotations to be low-resolution.  Arguably it might be better to
preserve the original time scale of each annotator.

Also note that the output afreq equals the input high-resolution sfreq
- which might not be the same as the output high-resolution sfreq, if
not all signals are copied!