xform: set correct sampling frequency for records with no signals.
As far as I can tell, the setsampfreq(0) is useless except where it's
a bug.  If there are any input signals, then the next thing we do is
to call isigopen again for the same record.  If an output header file
is used, we call sampfreq (which calls setsampfreq(0) itself) and
restore the original sampling frequency afterwards.

If there are no input signals and no output header file, then setting
the sampling frequency to 0 breaks the parsing of start/end times (and
also breaks conversion of annotation files with an explicit time
resolution.)