wfdb_free_config: unset environment variables before freeing.
Nothing says that it's okay to free strings that are part of the
environment just because the program is about to exit.

This entire function is arguably a bit silly - on "real" operating
systems, there is never any need to free memory upon exiting - but it
is slightly useful to the extent that it can reduce the number of
false positives reported by valgrind.