Fix deskewing for variable-layout records.
sigmap is modified to take the input frame as an argument, and
getskewedframe is modified to return a virtual frame rather than an
input frame.

Thus dsbuf is filled with virtual frames, so deskewing will work
correctly across variable-layout segment boundaries.

Furthermore, the actual deskewing is based on vsd, not isd, and
vsd[]->skew is initialized by make_vsd (when the first segment is
opened), so that the skews set in the first or layout segment override
any skews specified by later segments.

A side effect of this is that it will make wfdbgetskew work as
documented, and will make wfdbsetiskew work to a limited extent.  (I'm
not sure how long it's been since those functions actually did
anything, if they ever did.)