fortran: general improvements to string parameter handling.
In Fortran, when a string (CHARACTER*n) argument is passed to a
function or subroutine, the length of the buffer is passed as a hidden
extra argument.

For most older Fortran compilers, this extra argument is an 'int',
although newer versions of gfortran use 'size_t' instead.[1]  The
strings in question should be much smaller than INT_MAX, and on
currently-supported platforms, it should be safe to interpret a
'size_t' argument as an 'int' (but not the reverse); so we use 'int'
by default.  If necessary, the macro STRINGSIZE may be redefined in
order to use a different type.

If the macro NOSTRINGSIZE is defined, the length argument (if any) is
ignored.  Thus, wfdbf.c can now be compiled in several different
modes:

- NOSTRINGSIZE not defined, FIXSTRINGS defined: input strings are
  terminated following the last non-space character.  Output strings
  will be truncated if necessary, and padded with spaces to the full
  size of the buffer.  This is the traditional behavior for strings in
  Fortran.

- NOSTRINGSIZE not defined, FIXSTRINGS not defined: input strings are
  terminated by a zero or by the end of the buffer.  Output strings
  will always be zero-terminated, and will be truncated if necessary
  to fit in the buffer.

- NOSTRINGSIZE defined, FIXSTRINGS defined: historical (dangerous)
  behavior.  Input strings are terminated by the first space
  character.  Output strings will always be terminated by a single
  space character, and may overrun the buffer.

- NOSTRINGSIZE defined, FIXSTRINGS not defined: historical (dangerous)
  behavior.  Input strings are terminated by a zero.  Output strings
  will always be zero-terminated, and may overrun the buffer.

Thus, it is now possible for Fortran programs (using FIXSTRINGS) to
handle:

- time strings containing spaces (timstr, mstimstr, strtim, setbasetime)

- signal names containing spaces (getsiginfo, setsiginfo, findsig)

- info strings containing spaces (getinfo, putinfo)

- annotation aux values containing spaces (getann, putann)

Input strings are always copied into an internal buffer.  This has a
minor side effect on functions that take a "record" parameter
(annopen, isigopen, osigopen, wfdbinit, getinfo, setinfo, newheader,
setheader, sampfreq, wfdbfile): previously, if FIXSTRINGS was not
defined and the record name ended with ".hea", the "." would be
replaced by a null character.  Now, the input string is never
modified.

Finally, the internal functions fcstring and cfstring are not
externally visible.

[1] https://gcc.gnu.org/onlinedocs/gcc-9.2.0/gfortran/Argument-passing-conventions.html