fortran: setsiginfo: handle string arguments consistently.
The fname, desc, and units parameters are now copied into internal
storage.  If FIXSTRINGS is defined, trailing spaces are removed.

(Previously, the caller would be required to maintain these strings -
separately for each signal, if necessary - until after invoking
osigfopen or setheader.  The strings were also required to be
null-terminated.)