fortran: getann, ungetann: handle aux parameter.
In getann, the aux parameter is set to the annotation "aux" string, if
any.  (Previously, the parameter was unused.)

In ungetann, the aux parameter is interpreted in the same manner as
putann.  (Previously, the parameter was used as raw aux data; in
practice, this didn't matter because getann didn't handle aux.)