osig*open unification: split into four functions.
osigopen and osigfopen share a considerable amount of code.  To make
this more maintainable, the common logic needs to be moved into a
separate helper function.  To avoid inadvertently making functional
changes, this will be done in several patches.

This patch is the first step: move the core of osigopen into a new
function 'openosig', and move the core of osigfopen into a new
function 'openosig2'.  Eventually these two helper functions will be
identical and the second copy can be removed.