www_get_url_range_chunk: don't display network errors.
Instead, the caller should display an error message if the request
fails unexpectedly.  This was already done in all cases but one.

(If we are opening a new file, failures should generally be silent,
and simply result in wfdb_fopen returning NULL.  If we are reading
from a previously-opened file, failures are generally unexpected and
should be reported.)