nf_new, www_get_url_range_chunk: better use of range requests.
This should significantly improve the performance of opening a remote
file by eliminating unnecessary round-trips.

Instead of first requesting the file size and then requesting a byte
range, we can start by "blindly" requesting the first N bytes of the
file.  If the range we request is larger than the actual file length
(or if the server does not support retrieval of byte ranges for the
given file), the server will simply send the entire file.

Thus, if the server gives a response that is longer or shorter than
the range we request, we can assume that it is the entire file.  In
that case, we want to keep the entire file in memory (force it to
NF_FULL_MODE).

(Previously, if the response was longer than the requested range,
www_get_url_range_chunk would truncate it, meaning that applications
would typically request the same data over and over.)

If the first request succeeds but the file size is still unknown, then
send a second request to determine the size of the file.

Note that this maintains the (historical, although unintuitive)
behavior of treating empty files as nonexistent.