www_get_*: don't close the connection on HTTP errors.
CURLOPT_FAILONERROR means curl will close the connection as soon as it
sees a 4xx or 5xx status line.  This is not good because we often
request multiple nonexistent URLs, especially when starting up, and
opening a new connection is slow.

If we don't use FAILONERROR, we need to check the response code after
every request (and assume that all future protocols will maintain the
convention that codes up to 399 are "success" and 400 and above are
"failure".)  We also need to be careful not to treat the contents of
the error document, if any, as valid data.