www_init: follow HTTP redirections.
It's worth noting that previous libcurl-based versions of libwfdb have
never correctly handled HTTP redirections; they simply treated the
error page as an ordinary file.  (libwww-based versions of libwfdb, in
contrast, did handle redirections properly.)

Consequently, we add a flag ("+3XX") to the User-Agent string, so that
the server can easily detect whether the client will understand
redirections.  In the future, the PhysioNet servers may temporarily
redirect clients if this flag is present and a faster mirror is
available.

Note that the username and password are determined (based on
WFDBPASSWORD) for the initial URL only.  If we are redirected to
another location on the same host, libcurl will keep the same username
and password; if we are redirected to another host, libcurl will not
send any username or password.  Servers that provide access-controlled
data must take this into account, and assume that username and
password *may or may not* be available following a redirection.

The URL returned by wfdbfile() is the original URL, regardless of
whether we have been redirected.  Likewise, only the original URL's
directory name is added to the temporary database path.