nf_new, nf_get_range: cache HTTP redirections.
Following a redirection, subsequent requests for the same WFDB_FILE
will use the new URL, for up to five minutes after the initial
request.

(This only matters for NF_CHUNK_MODE and www_get_url_range_chunk; in
NF_FULL_MODE, the entire file is cached, so there are never any
subsequent requests.)

In theory, it would be better to take into account the HTTP response
codes (permanent versus temporary) and cache-control headers, if any,
but this could be extremely complicated when dealing with multiple
levels of redirection.