isigopen: limit total spf when switching segments.
Various functions can be made to misbehave if the frame size (total
spf) of a segment is greater than that of the layout header.

There are two primary sub-issues:

 - Because getframe uses the caller's vector as a buffer for the
   segment frame (and subsequently calls sigmap to rearrange it), it
   will overflow the caller's buffer if the segment frame is larger
   than the virtual frame.

 - Because isigopen itself reallocates tvector, and getvec may
   indirectly invoke isigopen, we may end up writing into a buffer
   after it has been freed.  Similar problems would occur with uvector
   and isgsettime, and with vvector and tnextvec.

For fixed-layout records, require every signal to have exactly the
expected number of samples per frame; the results otherwise will be
nonsensical.

For variable-layout records, we still tolerate signals with the wrong
number of samples per frame, since sigmap will rearrange the result
into something that is garbled but still (somewhat) meaningful.

Limiting the frame size is not the best solution; it would be better
to fix the underlying problems with getskewedframe/sigmap and with the
reallocation of tvector/uvector/vvector.  However, this should work to
fix the immediate problems.

Note that 'tspf' is now set in all cases (fixed-layout and
single-segment as well as variable-layout.)

'framelen', in contrast, is consistently set according to the current
set of input signals.

Note that if sigmap_init fails (which was not previously possible),
isigopen will now close all input signals and return -3.  This should
not be relevant to applications unless they attempt to open a new
record, after opening a multi-segment record and without first calling
wfdbquit(), which unfortunately has never really worked correctly.