getskewedframe: error if nisig = 0 and nvsig > 0.
This means that we are reading a non-trivial multi-segment record
(either fixed layout or variable layout), and we failed to open the
current segment for some reason.

The only way this can happen is if the previous call to isigopen
(either directly or via getframe or isgsettime) failed, and those
errors should be reported back to the application.

However, if the error is not handled for some reason, but the
application expects that some positive number of signals are still
open, we should not return 0.